new_requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.2.1
django-bootstrap-datepicker-plus==5.0.5
django-crispy-forms==2.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
